classification-service/requirements.txt
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.54.0
Deprecated==1.2.18
docker==7.1.0
fastapi==0.115.12
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.3
Mako==1.3.10
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.0
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==6.31.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.3.0
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
